Add files via upload
This group project shows how to connect the main program in Python to a local instance of SQL database. My team and I built a mock dataset with library attendance data (for a lack of official data that university library could not share with us). The project started in August 2020, during the Covid-19 pandemic and was intended to inform students who wanted to study at the school library of how crowded (and thereby Covid-risky) it can become.

The program allows users to enter their student ID, age, sex, and day and time they want to come to the library. Then, the program checks with the SQL database (filled with our mock data) how populated the library has been during those days/hours and shows a pie chart to the user with 4 time slots for that day and the percentages of library 'crowdedness' for each time slot. The idea was to essentially warn the user that if they go to the library, which is indoors and can become crowded, they are exposing themselves to the risk of getting Covid-19 to a certain extent.

My tasks in this project included building the input data validation functions and an integration of the program with the local instance of SQL database.

Date: 11-30-2020